# Make permutations type-stable