For some reason, Julia can't infer the return type of the `next` method for `permutations`. This is just a very small change that fixes this issue. 

Before:
```
julia> @benchmark collect(permutations(1:10, 3))
BenchmarkTools.Trial: 
  memory estimate:  433.48 kb
  allocs estimate:  10085
  --------------
  minimum time:     749.489 μs (0.00% GC)
  median time:      770.428 μs (0.00% GC)
  mean time:        817.017 μs (4.68% GC)
  maximum time:     3.858 ms (78.25% GC)
  --------------
  samples:          6117
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

After:
```
julia> @benchmark collect(permutations(1:10, 3))
BenchmarkTools.Trial: 
  memory estimate:  275.95 kb
  allocs estimate:  4324
  --------------
  minimum time:     90.095 μs (0.00% GC)
  median time:      97.903 μs (0.00% GC)
  mean time:        121.701 μs (17.90% GC)
  maximum time:     3.727 ms (95.66% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```