@ti-s 2016-11-25
Superseded by https://github.com/JuliaMath/Combinatorics.jl/pull/37 (except the single inference test).